fix panic!

+++ barrios.go
@@ -14,14 +14,14 @@ func (network *Network) neighbourhoodMaintenance() {
 		observation := network.local.getObservation(name)
 		vibe := network.calculateVibe(name)
 
-		clusterIndex := vibe % len(clusterNames)
+		clusterIndex := vibe % uint64(len(clusterNames))
 		observation.ClusterName = clusterNames[clusterIndex]
 		observation.ClusterEmoji = BarrioEmoji[clusterIndex]
 		network.local.setObservation(name, observation)
 	}
 }
 
-func (network *Network) calculateVibe(name string) int {
+func (network *Network) calculateVibe(name string) uint64 {
 	// vibe is based on the name and the current month
 	// so neighbourhoods shift over time but stay consistent across the network
 	hasher := sha256.New()
@@ -35,5 +35,5 @@ func (network *Network) calculateVibe(name string) int {
 	// use the first 8 bytes of hash as a uint64
 	vibe := binary.BigEndian.Uint64(hash[:8])
 
-	return int(vibe)
+	return vibe
 }
